requirements.txt
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.2.0
chevron==0.14.0
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.103.1
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
idna==3.4
loguru==0.7.2
msgpack==1.0.5
multidict==6.0.4
nonebot-adapter-onebot==2.2.4
nonebot-plugin-gocqhttp==0.6.12
nonebot2==2.1.0
psutil==5.9.5
py-cpuinfo==9.0.0
pydantic==1.10.12
pygtrie==2.5.0
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.0.4
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
yarl==1.9.2
